requirements.txt
Flask
flask-session
PyMuPDF
python-docx
google-generativeai
